+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - [BREAKING] `CheckNullifiersByPrefix` now takes a starting block number (#707).
 - [BREAKING] Removed nullifiers from `SyncState` endpoint (#708).
 - [BREAKING] Added `batch_prover_url` to block producer configuration (#701).
+- [BREAKING] Added `block_prover_url` to block producer configuration (#719).
 
 ### Enhancements